+++ ui/src/win/label.cpp
@@ -39,7 +39,7 @@ namespace xaml
             this->__create(params);
         }
         rectangle real = region - get_margin();
-        THROW_IF_WIN32_BOOL_FALSE(SetWindowPos(get_handle(), HWND_TOP, real.x, real.y, real.width, real.height, SWP_NOZORDER));
+        THROW_IF_WIN32_BOOL_FALSE(SetWindowPos(get_handle(), HWND_TOP, (int)real.x, (int)real.y, (int)real.width, (int)real.height, SWP_NOZORDER));
         __set_size_noevent({ real.width, real.height });
         draw_text();
         draw_alignment();
@@ -49,7 +49,7 @@ namespace xaml
 
     void label::draw_size()
     {
-        THROW_IF_WIN32_BOOL_FALSE(SetWindowPos(get_handle(), HWND_TOP, 0, 0, get_width(), get_height(), SWP_NOZORDER | SWP_NOMOVE));
+        THROW_IF_WIN32_BOOL_FALSE(SetWindowPos(get_handle(), HWND_TOP, 0, 0, (int)get_width(), (int)get_height(), SWP_NOZORDER | SWP_NOMOVE));
     }
 
     void label::draw_text()
